fix: move up GetRightsRequest to avoid multiple useless requests
